Split peering models in several file
Also get rid of `TemplateModel`, stop using `to_dict()`.
