+++ utils/models.py
@@ -133,49 +133,3 @@ class TaggableModel(models.Model):
 
     class Meta:
         abstract = True
-
-
-class TemplateModel(models.Model):
-    """
-    Abstract class providing functions to be used in templates.
-    """
-
-    class Meta:
-        abstract = True
-
-    def to_dict(self, ignore=[]):
-        data = {}
-
-        # Iterate over croncrete and many-to-many fields
-        for field in self._meta.concrete_fields + self._meta.many_to_many:
-            if field.name in ignore:
-                continue
-
-            value = None
-
-            # If the value of the field is another model, fetch an instance of it
-            # Exception made of tags for which we just retrieve the list of them for later
-            # conversion to simple strings
-            if isinstance(field, ForeignKey):
-                value = self.__getattribute__(field.name)
-            elif isinstance(field, ManyToManyField):
-                value = list(self.__getattribute__(field.name).all())
-            elif isinstance(field, TaggableManager):
-                value = list(self.__getattribute__(field.name).all())
-            else:
-                value = self.__getattribute__(field.name)
-
-            # If the instance of a model as a to_dict() function, call it
-            if isinstance(value, TemplateModel):
-                data[field.name] = value.to_dict(ignore=ignore)
-            elif isinstance(value, list):
-                data[field.name] = []
-                for element in value:
-                    if isinstance(element, TemplateModel):
-                        data[field.name].append(element.to_dict(ignore=ignore))
-                    else:
-                        data[field.name].append(str(element))
-            else:
-                data[field.name] = value
-
-        return data
